Refactor: Modularization
