+++ json_portfolio/json_portfolio.py
@@ -1,234 +1,7 @@
 import reflex as rx
-import json
 
-from pathlib import Path
+from .components import *
 from .styles import *
-from typing import Any
-
-
-def load_portfolio() -> dict[str, Any]:
-    """Load portfolio data from the resume.json file."""
-    with open(Path(__file__).parent / "resume.json", "r", encoding="utf-8") as f:
-        return json.load(f)
-
-
-data = load_portfolio()
-
-
-class ContactState(rx.State):
-    show_modal: bool = False
-
-    def toggle_modal(self):
-        self.show_modal = not self.show_modal
-
-
-def navbar() -> rx.Component:
-    return rx.hstack(
-        rx.hstack(
-            rx.link("Home", href="/", **link_style),
-            rx.link("Work", href="#work", **link_style),
-            rx.link("Education", href="#education", **link_style),
-            rx.link("Projects", href="#projects", **link_style),
-            **nav_links_container
-        ),
-
-        rx.spacer(),
-
-        rx.hstack(
-            rx.link(
-                rx.image(src="/github.svg", **icon_image_size),
-                is_external=True
-            ),
-            rx.link(
-                rx.image(src="/linkedin.svg", **icon_image_size),
-                is_external=True
-            ),
-            **nav_icons_container
-        ),
-
-        **navbar_style
-    )
-
-
-def contact_modal() -> rx.Component:
-    basics = data["basics"]
-    return rx.cond(
-        ContactState.show_modal,
-        rx.box(
-            rx.box(
-                rx.heading("Contact Information", **section_heading_style),
-
-                rx.vstack(
-                    rx.hstack(
-                        rx.icon(tag="mail"),
-                        rx.text(basics["email"]),
-                        align_items="center"
-                    ),
-                    rx.hstack(
-                        rx.icon(tag="phone"),
-                        rx.text(basics.get("phone", "N/A")),
-                        align_items="center"
-                    ),
-                    rx.hstack(
-                        rx.icon(tag="map-pin"),
-                        rx.text(
-                            f"{basics['location']['city']}, "
-                            f"{basics['location']['region']}, "
-                            f"{basics['location']['countryCode']}"
-                        ),
-                        align_items="center"
-                    ),
-                    **contact_info_style
-                ),
-
-                rx.center(
-                    rx.button("Close", on_click=ContactState.toggle_modal, **modal_button_style),
-                    margin_top="16px"
-                ),
-
-                **modal_card_style
-            ),
-            **modal_overlay_style
-        ),
-        None
-    )
-
-
-def header_section() -> rx.Component:
-    """Render header with name/title aligned vertically center with image."""
-    basics = data["basics"]
-
-    return rx.vstack(
-        rx.grid(
-            rx.vstack(
-                rx.heading(basics["name"], size="8"),
-                rx.text(basics["label"], font_weight="bold"),
-                rx.button("Contact", on_click=ContactState.toggle_modal, **contact_button_style),
-                **header_text_block
-            ),
-            rx.box(
-                rx.image(src=basics["image"], **header_image_style),
-                display="flex",
-                justify_content="flex-end"
-            ),
-            **header_grid_style
-        ),
-        rx.text(basics["summary"], **summary_text_style),
-        **header_section_style
-    )
-
-
-def education_section() -> rx.Component:
-    """Render the education section."""
-    return rx.vstack(
-        *[
-            rx.box(
-                rx.heading(
-                    f"{edu['studyType']} in {edu['area']}",
-                    **heading_education_style
-                ),
-                rx.text(edu["institution"]),
-                rx.text(f"{edu['startDate']} – {edu.get('endDate') or 'Present'}"),
-                rx.text(f"GPA: {edu['score']}") if edu.get("score") else None,
-                rx.unordered_list(
-                    *[rx.list_item(course) for course in edu.get("courses", [])]
-                ) if edu.get("courses") else None,
-                **education_item_style
-            )
-            for edu in data["education"]
-        ],
-        **education_section_style
-    )
-
-
-def certificates_section() -> rx.Component:
-    """Render the certificates section."""
-    return rx.vstack(
-        *[
-            rx.box(
-                rx.heading(cert["name"], **certificate_heading_style),
-                rx.text(f"Issued by {cert['issuer']} on {cert['date']}"),
-                rx.link(
-                    "View Certificate",
-                    href=cert["url"],
-                    is_external=True
-                ) if cert.get("url") else None,
-                **certificate_item_style
-            )
-            for cert in data["certificates"]
-        ],
-        **certificates_section_style
-    )
-
-
-def work_section() -> rx.Component:
-    """Render the work experience section."""
-    return rx.vstack(
-        *[
-            rx.box(
-                rx.heading(f'{job["position"]} @ {job["name"]}', **work_heading_style),
-                rx.text(
-                    f'{job["startDate"]} – '
-                    f'{job.get("endDate") or "Present"}'
-                ),
-                rx.text(job["summary"]),
-                rx.unordered_list(
-                    *[rx.list_item(item) for item in job["highlights"]]
-                ),
-                **work_item_style
-            )
-            for job in data["work"]
-        ],
-        **work_section_style
-    )
-
-
-def project_section() -> rx.Component:
-    """Render the projects section in a grid of cards."""
-    return rx.box(
-        rx.grid(
-            *[
-                rx.box(
-                    rx.vstack(
-                        rx.heading(project["name"], **project_heading_style),
-                        rx.text(project["description"]),
-
-                        rx.hstack(
-                            *[
-                                rx.box(
-                                    tech,
-                                    **tech_tag_style
-                                )
-                                for tech in project["highlights"]
-                            ],
-                            **tech_tag_container_style
-                        ),
-
-                        rx.text(project["role"], font_weight="bold"),
-
-                        rx.link(
-                            "GitHub",
-                            href=project["github"],
-                            is_external=True
-                        ) if project.get("github") else None,
-
-                        **project_content_style
-                    ),
-                    **project_card_style
-                )
-                for project in data["projects"]
-            ],
-            **project_grid_style
-        )
-    )
-
-
-def footer() -> rx.Component:
-    """Render footer."""
-    return rx.vstack(
-        rx.text("© 2025 Matias Falconaro.", **footer_text_style),
-        **footer_style
-    )
 
 
 def index() -> rx.Component:
@@ -238,19 +11,27 @@ def index() -> rx.Component:
             header_section(),
             rx.divider(),
 
-            rx.heading("Work Experience", id="work", **section_heading_style),
+            rx.heading("Work Experience",
+                       id="work",
+                       **section_heading_style),
             work_section(),
             rx.divider(),
 
-            rx.heading("Education", id="education", **section_heading_style),
+            rx.heading("Education",
+                       id="education",
+                       **section_heading_style),
             education_section(),
             rx.divider(),
 
-            rx.heading("Certificates", id="certificates", **section_heading_style),
+            rx.heading("Certificates",
+                       id="certificates",
+                       **section_heading_style),
             certificates_section(),
             rx.divider(),
 
-            rx.heading("Projects", id="projects", **section_heading_style),
+            rx.heading("Projects",
+                       id="projects",
+                       **section_heading_style),
             project_section(),
             rx.divider(),
